Fix under/overflow issues in TiffReader.
Relates to drewnoakes/metadata-extractor#167

+++ MetadataExtractor/Formats/Exif/ExifTiffHandler.cs
@@ -190,7 +190,7 @@ namespace MetadataExtractor.Formats.Exif
             return false;
         }
 
-        public override bool TryCustomProcessFormat(int tagId, TiffDataFormatCode formatCode, int componentCount, out int byteCount)
+        public override bool TryCustomProcessFormat(int tagId, TiffDataFormatCode formatCode, uint componentCount, out long byteCount)
         {
             if ((ushort)formatCode == 13u)
             {
